# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A painless HTML to PDF rendering service. Generate PDF reports and documents from HTML templates or raw HTML.",
        "title": "PdfTurtle API",
        "contact": {
            "name": "Lucas Gaitzsch",
            "email": "lucas@gaitzsch.dev"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://github.com/lucas-gaitzsch/pdf-turtle/blob/main/LICENSE"
        },
        "version": "1.1"
    },
    "paths": {
        "/api/html-bundle": {
            "get": {
                "description": "List HTML bundles from server, allowing to render PDFs from it at a later time",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List HTML-Bundles Info"
                ],
                "summary": "List HTML bundles from server",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Save HTML bundle to server, allowing to render PDFs from it at a later time",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Save HTML-Bundle"
                ],
                "summary": "Save HTML bundle to server",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Bundle Zip-File",
                        "name": "bundle",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the bundle",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the bundle",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Template engine to use for template",
                        "name": "templateEngine",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/html-bundle/{id}": {
            "get": {
                "description": "Get HTML bundle from server, allowing to render PDFs from it at a later time",
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Get HTML-Bundle"
                ],
                "summary": "Get HTML bundle from server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the bundle",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}